requirements.txt
django-ckeditor
pillow
pytils
django-colorfield
